pyproject.toml
[tool.poetry]
name = "lifecycle-platform"
version = "0.1.0"
description = "Lifecycle platform"
authors = ["Qualitara"]

[tool.poetry.dependencies]
python = "^3.12"
google-cloud-bigquery = "^3.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
ruff = "^0.11"
apache-airflow = "^2.10.0"

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "RUF012", # mutable class attrs — false positive on constant lookup dicts
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"